@@ -1,12 +1,13 @@
 package neil_opena.cyphergazer;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class EncryptionFragment extends Fragment{
+public class EncryptionFragment extends Fragment {
 
     public static EncryptionFragment newInstance(){
         return new EncryptionFragment();
